Update SellValue.lua
Fixed error that occurs in combination with EQL3. The error specifically happens when you open the quest log, then mouse over a creature (or maybe other object creating a info window) and then mousing over the quest log item again.
The quest log item does not produce a box with item information, which causes the links to fail.

The underlying bug is highly likely hidden somewhere inside EQL3, but I do not know how to find it there.